requirements.txt
openai==1.59.6
langchain== 0.2.8
langchain_openai==0.1.16
langchain-core==0.2.22
python-dotenv==1.0.1
unstructured == 0.14.10
langgraph==0.1.8
langchain-openai == 0.1.16
langchain-community==0.2.7
langchain-text-splitters==0.2.2
pydantic==2.7.4
pydantic-settings==2.9.1
python-multipart==0.0.9 
pandas== 2.3.3
openpyxl>=3.1.0
scikit-learn>=1.3.0
xgboost==1.7.6            
lightgbm>=4.0.0            
imbalanced-learn>=0.11.0     
shap>=0.43.0
eli5>=0.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
folium>=0.14.0          
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
scipy>=1.11.0
statsmodels>=0.14.0
tqdm>=4.66.0
joblib>=1.3.0
geopy>=2.4.0
requests>=2.31.0